обертка над моделью
